\name{IPBT_rank-package}
\alias{IPBT_rank-package}
\alias{IPBT_rank}
\docType{package}
\title{
\packageTitle{IPBT_rank}
}
\description{
\packageDescription{IPBT_rank}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IPBT_rank}
\packageIndices{IPBT_rank}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{IPBT_rank}

Maintainer: \packageMaintainer{IPBT_rank}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
