\name{rankIPBT-package}
\alias{rankIPBT-package}
\alias{rankIPBT}
\docType{package}
\title{
\packageTitle{rankIPBT}
}
\description{
\packageDescription{rankIPBT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rankIPBT}
\packageIndices{rankIPBT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rankIPBT}

Maintainer: \packageMaintainer{rankIPBT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
